[refactory] model, service 패키지 리팩토링
