+++ PayStory/src/main/java/com/AourZ/PayStory/controller/AIController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.AourZ.PayStory.model.ExpenditureVO;
+import com.AourZ.PayStory.model.accountBook.ExpenditureVO;
 import com.AourZ.PayStory.service.ai.ChatbotService;
 import com.AourZ.PayStory.service.ai.OCRService;
 import com.AourZ.PayStory.service.ai.STTService;
