+++ PayStory/src/main/java/com/AourZ/PayStory/controller/BoardController.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.AourZ.PayStory.model.BoardVO;
+import com.AourZ.PayStory.model.board.BoardVO;
 import com.AourZ.PayStory.service.BoardService;
 
 @Controller
