+++ PayStory/src/main/java/com/AourZ/PayStory/dao/IShareAccountDAO.java
@@ -1,8 +1,8 @@
 package com.AourZ.PayStory.dao;
 
-import com.AourZ.PayStory.model.AccountBookVO;
-import com.AourZ.PayStory.model.ShareAccountBookVO;
-import com.AourZ.PayStory.model.ShareBudgetVO;
+import com.AourZ.PayStory.model.accountBook.AccountBookVO;
+import com.AourZ.PayStory.model.accountBook.ShareAccountBookVO;
+import com.AourZ.PayStory.model.accountBook.ShareBudgetVO;
 
 public interface IShareAccountDAO {
 	/* ArrayList<ProductVO> listAllProduct(); */  		// 가계부 조회
