+++ PayStory/src/main/java/com/AourZ/PayStory/dao/MemberDAO.java
@@ -10,10 +10,10 @@ import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import com.AourZ.PayStory.model.BoardVO;
-import com.AourZ.PayStory.model.LoginVO;
-import com.AourZ.PayStory.model.MemberVO;
-import com.AourZ.PayStory.service.AccountBookService;
+import com.AourZ.PayStory.model.board.BoardVO;
+import com.AourZ.PayStory.model.member.LoginVO;
+import com.AourZ.PayStory.model.member.MemberVO;
+import com.AourZ.PayStory.service.accountBook.AccountBookService;
 
 @Repository
 public class MemberDAO implements IMemberDAO {
