+++ PayStory/src/main/java/com/AourZ/PayStory/model/accountBook/EditorVO.java
@@ -1,4 +1,4 @@
-package com.AourZ.PayStory.model;
+package com.AourZ.PayStory.model.accountBook;
 
 public class EditorVO {
 	private String editDate;		// 수정 날짜
