+++ PayStory/src/main/java/com/AourZ/PayStory/model/accountBook/ShareBudgetVO.java
@@ -1,4 +1,4 @@
-package com.AourZ.PayStory.model;
+package com.AourZ.PayStory.model.accountBook;
 
 public class ShareBudgetVO {
