+++ PayStory/src/main/java/com/AourZ/PayStory/model/accountBook/TagTotalVO.java
@@ -1,4 +1,4 @@
-package com.AourZ.PayStory.model;
+package com.AourZ.PayStory.model.accountBook;
 
 public class TagTotalVO {
 	private String date;	// 날짜 (년 / 년-월 / 년-월-일)
