+++ PayStory/src/main/java/com/AourZ/PayStory/service/IMemberService.java
@@ -2,9 +2,9 @@ package com.AourZ.PayStory.service;
 
 import java.util.List;
 
-import com.AourZ.PayStory.model.BoardVO;
-import com.AourZ.PayStory.model.LoginVO;
-import com.AourZ.PayStory.model.MemberVO;
+import com.AourZ.PayStory.model.board.BoardVO;
+import com.AourZ.PayStory.model.member.LoginVO;
+import com.AourZ.PayStory.model.member.MemberVO;
 
 public interface IMemberService {
 	public void register(MemberVO memberVO) throws Exception;
